// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  USER: 'USER',
  ADMIN: 'ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const Status = {
  PENDING: 'PENDING',
  VERIFIED: 'VERIFIED'
} as const

export type Status = (typeof Status)[keyof typeof Status]


export const EntryType = {
  CREDIT: 'CREDIT',
  DEBIT: 'DEBIT'
} as const

export type EntryType = (typeof EntryType)[keyof typeof EntryType]


export const Category = {
  REBATE: 'REBATE',
  WITHDRAWAL: 'WITHDRAWAL',
  FEE: 'FEE'
} as const

export type Category = (typeof Category)[keyof typeof Category]


export const TicketType = {
  WITHDRAWAL: 'WITHDRAWAL',
  VERIFICATION: 'VERIFICATION'
} as const

export type TicketType = (typeof TicketType)[keyof typeof TicketType]


export const TicketStatus = {
  PENDING: 'PENDING',
  DONE: 'DONE'
} as const

export type TicketStatus = (typeof TicketStatus)[keyof typeof TicketStatus]


export const ActionTokenType = {
  EMAIL_VERIFICATION: 'EMAIL_VERIFICATION',
  PASSWORD_RESET: 'PASSWORD_RESET'
} as const

export type ActionTokenType = (typeof ActionTokenType)[keyof typeof ActionTokenType]